[core] Migrate EventEmitter
